requirements.txt
# Core Django
Django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
django-mptt==0.15.0
django-postgres-metrics==0.7.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-otp==1.2.4
cryptography==41.0.7
PyJWT==2.8.0

# Caching & Performance
redis==5.0.1
django-redis==5.4.0
django-cachalot==2.5.1

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
kombu==5.3.4

# File Storage & Media
boto3==1.34.0
django-storages==1.14.2
Pillow==10.1.0

# API & GraphQL
graphene-django==3.1.5
django-filter==23.5
drf-spectacular==0.26.5

# Data Processing & Analytics
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
prophet==1.1.5
tensorflow==2.15.0

# Financial & Accounting
python-dateutil==2.8.2
babel==2.13.1
forex-python==1.8

# Banking Integrations
requests==2.31.0
lxml==4.9.3

# Testing & Quality
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0